Refactor models
- move models
- also fix spelling and missing types
