+++ public/models/convertFormat.model.ts
@@ -1,0 +1,16 @@
+//Used in the convert functions
+
+export interface ConvertProject {
+  name: string;
+  fullPath: string;
+  books: ConvertBook[];
+}
+
+export interface ConvertBook {
+  name: string;
+  chapters: ConvertChapter[];
+}
+
+export interface ConvertChapter {
+  name: string;
+}
