+++ public/models/projectData.model.ts
@@ -1,0 +1,3 @@
+export interface ProjectData {
+  readonly outputLocation: string;
+}
