+++ public/models/timings.model.ts
@@ -1,0 +1,19 @@
+export type Timings = LineTiming[];
+
+// this follows the spec ... ?
+interface LineTiming {
+  readonly type: 'caption';
+  readonly index: number;
+  readonly start: number;
+  readonly end: number;
+  readonly duration: number;
+  readonly content: string;
+  readonly text: string; // not used
+  readonly words: WordTiming[];
+}
+
+interface WordTiming {
+  readonly word: string;
+  readonly start: number;
+  readonly end: number;
+}
